requirements.txt
# Core Data Science
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# NFL Data
nfl-data-py>=0.3.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# API & Web
requests>=2.31.0
aiohttp>=3.8.0

# Configuration
python-dotenv>=1.0.0

# Jupyter (for notebooks)
jupyter>=1.0.0
ipykernel>=6.25.0

# Data Validation
pandera>=0.17.0

# Progress bars
tqdm>=4.65.0

